requirements.txt
nemo_toolkit
streamlit

